requirements.txt
pandas
faiss-cpu
sentence-transformers
requests
python-dotenv
